+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [0.1.10] (2021-05-28)
+## [0.1.11](https://github.com/SIMBAChain/libsimba4j-platform/compare/v0.1.10...v0.1.11) (2021-05-23)
 
+### Features
+
+* Add get transaction count
+* Add get server side wallet
+
+## [0.1.10] (2021-03-28)
 
 ### Features